data validation class moved to another module.
